feat: add integration test for the stakeholder
